[EC-73] feat: move collection dialog to shared
