+++ apps/web/src/app/organizations/organization.module.ts
@@ -8,7 +8,6 @@ import { AccessSelectorModule } from "./components/access-selector";
 import { GroupAddEditComponent } from "./manage/group-add-edit.component";
 import { GroupsComponent } from "./manage/groups.component";
 import { UserGroupsComponent } from "./manage/user-groups.component";
-import { CollectionDialogModule } from "./components/collection-dialog";
 import { OrganizationsRoutingModule } from "./organization-routing.module";
 import { SharedOrganizationModule } from "./shared";
 import { GroupServiceAbstraction } from "./services/abstractions/group";
@@ -17,10 +16,9 @@ import { GroupService } from "./services/group/group.service";
 @NgModule({
   imports: [
     SharedModule,
-    AccessSelectorModule,
-    CollectionDialogModule,
     OrganizationsRoutingModule,
     SharedOrganizationModule,
+    AccessSelectorModule,
   ],
   declarations: [
     GroupsComponent,
