+++ apps/web/src/app/organizations/shared/components/collection-dialog/collection-dialog.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 
-import { SharedModule } from "../../../shared";
-import { AccessSelectorModule } from "../access-selector";
+import { SharedModule } from "../../../../shared";
+import { AccessSelectorModule } from "../../../components/access-selector";
 
 import { CollectionDialogComponent } from "./collection-dialog.component";
