+++ apps/web/src/app/organizations/shared/index.ts
@@ -1,1 +1,2 @@
 export * from "./shared-organization.module";
+export * from "./components/collection-dialog";
